Migración de la configuración de base de datos de MySQL a PostgreSQL en application.properties. Actualización de la contraseña y parámetros de conexión. Verificación exitosa de la conexión y generación automática de tablas por Spring Boot. Backend listo para desarrollo y pruebas sobre PostgreSQL.

+++ src/main/java/com/dresscode/api_dresscode/services/DescuentoService.java
@@ -36,7 +36,7 @@ public class DescuentoService extends BaseServiceImpl<Descuento, Long>{
         Descuento descuento = Descuento.builder()
                 .fechaInicio(descuentoDTO.getFechaInicio())
                 .fechaCierre(descuentoDTO.getFechaCierre())
-                .porcentajeDescuento(descuentoDTO.getPorcentajeDescuento())
+                .porcentajeDescuento(descuentoDTO.getPorcentajeDescuento().doubleValue())
                 .build();
 
         return descuentoRepository.save(descuento);
@@ -52,7 +52,7 @@ public class DescuentoService extends BaseServiceImpl<Descuento, Long>{
 
         descuentoExistente.setFechaInicio(datosActualizados.getFechaInicio());
         descuentoExistente.setFechaCierre(datosActualizados.getFechaCierre());
-        descuentoExistente.setPorcentajeDescuento(datosActualizados.getPorcentajeDescuento());
+        descuentoExistente.setPorcentajeDescuento(datosActualizados.getPorcentajeDescuento().doubleValue());
         descuentoExistente.setProductos(descuentoExistente.getProductos());
 
         return descuentoRepository.save(descuentoExistente);
